fixed document-downloader: download images from KoBo & save on OneDrive
